+++ docs/conf.py
@@ -239,7 +239,7 @@ intersphinx_mapping.update({
     'gsl': ('https://www.gnu.org/software/gsl/doc/html/', None),
     'gwcelery': ('https://gwcelery.readthedocs.io/en/latest/', None),
     'gwpy': ('https://gwpy.github.io/docs/latest/', None),
-    'ligo-lw': ('https://docs.ligo.org/kipp/python-ligo-lw/', None),
+    'igwn-ligolw': ('https://docs.ligo.org/computing/software/igwn-ligolw/', None),
     'pytest': ('https://docs.pytest.org/en/latest/', None),
     'reproject': ('https://reproject.readthedocs.io/en/stable/', None),
 })