+++ ligo/skymap/io/tests/test_io_events.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import re
 import subprocess
 
-from ligo.lw.utils import load_filename
+from igwn_ligolw.utils import load_filename
 import h5py
 import numpy as np
 import pytest
@@ -118,7 +118,7 @@ def test_sqlite(tmpdir):
     # Convert the test data to SQLite format in the temporary directory.
     xmlfilename = os.path.join(DATA_PATH, '2016_subset.xml.gz')
     dbfilename = str(tmpdir / '2016_subset.sqlite')
-    subprocess.check_call(['ligolw_sqlite', '-p',
+    subprocess.check_call(['igwn_ligolw_sqlite', '-p',
                            xmlfilename, '-d', dbfilename])
 
     # Symbolicly link the PSD directory into the temporary directory.